hip: remove unified-memory-only zero-memset unit test from Kokkos_CoreUnitTest_HIP
Wrapping the result of a system allocation into a HIPSpace view is not
valid Kokkos.
Even if it were, this test only works on MI300A when HSA_XNACK=1, which is
a user-controlled environment variable, not something Kokkos manages.
I guess a user is free to do this on MI300A if they want to run with
HSA_XNACK=1, but it's outside the Kokkos programming model.

Signed-off-by: Carl Pearson <cwpears@sandia.gov>