Disable bugprone-unused-raii in the Clang+CUDA build on ORNL Jenkins CI
The `CUDA-11.0.3-Clang-Tidy` has been failing since the merge of #7744

As it turns out, the `NOLINTBEGIN` and `NOLINTEND` commands were only
introduced in Clang-Tidy 14, whereas our CI job uses version 12.

Considering that the numeric traits is riddled with places that would
raise that warning, I propose to just wholesale diable it in builds that
use Clang-Tidy <14.