Fixing memory error in tuning API. The make_candidate_set double and int64_t implementations don't correctly copy the set of candidates, but just copy the pointer. This change will copy the set.
At some point, this code will need to add an explcit destructor
for the VariableValue struct, to clean up the memory allocated
when constructing the struct. These variables are only meant to
be declared once, so any currently leaked memory is small, but it
is a leak.

This fixes issue #7874