[[ci skip]] Sketch of how self similar interface could work
This represents the farthest I got via some experiments in sketching
out the self similar interface idea. I did more things, but this is
a condensed, and cut back version to convey the idea and the issue.

First of all: I came to the conclusion that a function like interface
- effectively a policy factory - works better than partial
  specialization for the execution policies.
The problem is that we need a mix of ctor args and template args for
policy. The latter don't necessarily have corresponding arguments for
the policy (think Schedule<Dynamic>). I guess we could introduce
dummy arguments if we'd like too, but its all pretty messy.
This approach avoids that completely via a function "range_policy".
From there we can use requires clauses to steer things properly:
we could return MDRangePolicy for multi dimensional ranges for
example.

The biggest issue is that I could for the live of me not figure
out how to make the "calling host function from device" function
go away. I tried a bunch of approaches (some of which are in the
code as comments). But it crashes in a CUDA build.

Anyway I wanted to document where I am and see if folks have
comments on the "policy factory" style used here.