Add AccessorArg_t View Ctor Property (#8112)
* Add AccessorArg_t View Ctor Property

This allows passing through a size_t from view construction to
the accessor. It is used as part of the View customization.
We later want to extent this to be templated, so we can pass in
other stuff than size_t, but I only need size_t for Sacado.

Signed-off-by: Christian Trott <crtrott@sandia.gov>

* Address Review comments

Signed-off-by: Christian Trott <crtrott@sandia.gov>

* Redo cmake logic for ViewSupport Test

Signed-off-by: Christian Trott <crtrott@sandia.gov>

* Work around MSVC+CUDA issue with multiple bases and pack expansion

Signed-off-by: Christian Trott <crtrott@sandia.gov>

* Disable View Customization Test for OpenACC

Fails with a similar issue as the one we had where passing data
handles by reference would result in illegal memory accesses.
I did the same here, but it still doesn't work.

Let this for someone else to figure out.

---------

Signed-off-by: Christian Trott <crtrott@sandia.gov>