Remove BUILD.md in favor of the documentation website
This file is bound to get out of date, we haven't touched it in 2y.
We have a guide discussing how to integrate Kokkos in a cmake project
and a guide on how to build from source.

The known issues with the Cray compiler and when linking with Fortran
needs to be migrated to the Known Issues in the doc if they are still
relevant.

The Spack information is redundant with the Spack.md file (which will
also be reviewed in a follow up).