Produce a diff when comparing kokkos/mdspan or desul/atomics
In https://github.com/kokkos/kokkos/actions/runs/18650866410/job/53168295770
the GitHub action log just says "Files A and B differ".
I think that it might be worth spitting out the actual diff at the risk
of printing more volume.  In the case of
https://github.com/kokkos/kokkos/pull/8562 it would have been usefu for
me to understand what was going on.

Signed-off-by: Damien L-G <dalg24@gmail.com>